idam_audit_log-master/requirements.txt
amqp==2.3.2
astroid==2.0.4
certifi==2018.10.15
chardet==3.0.4
colorama==0.4.0
dnspython==1.15.0
eventlet==0.24.1
greenlet==0.4.15
idna==2.7
importlib-metadata==0.6
isort==4.3.4
kombu==4.2.1
lazy-object-proxy==1.3.1
mccabe==0.6.1
mock==2.0.0
monotonic==1.5
nameko==2.12.0
path.py==11.5.0
pbr==5.1.0
pylint==2.1.1
python-arango==4.2.1
PyYAML==5.1.2
requests==2.20.0
six==1.11.0
typed-ast==1.1.0
urllib3==1.24
vine==1.1.4
Werkzeug==0.14.1
wrapt==1.10.11
